requirements.txt
pandas
openpyxl
ttkthemes
pillow